+++ src/gen_rules.mli
@@ -8,4 +8,4 @@ val gen
   -> Jbuild_load.conf
   -> (Build_interpret.Rule.t list *
      (* Evaluated jbuilds per context names *)
-     (Path.t * Pkgs.t * Stanzas.t) list String_map.t) Future.t
+     (Path.t * Scope.t * Stanzas.t) list String_map.t) Future.t